HTTP basic auth (#11)
This adds HTTP authentication to the data layer. Authentication is always enabled in production, where a `SHARED_SECRET` must be configured. It can also be enabled in development by generating a shared secret and adding it to `local.yml`.

We need username support during development but it's a few features away. I was on the fence about whether to add it now and ended up throwing it in. The way it works now, is anyone who knows the shared secret is trusted to identify themselves as a user. The username is [passed into Postgres](https://www.graphile.org/postgraphile/usage-library/#exposing-http-request-data-to-postgresql), where it can be used for authorization.

Frontend PR: curvewise/frontend#349

Ref https://www.wrike.com/open.htm?id=577538414